requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bleach==6.1.0
blinker==1.9.0
Cerberus==1.3.5
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.3.0
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
validators==0.22.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
